Completely avoid site.canonical blanking
The QR site needs the host, because there are no relative paths. But everything else was either the deploy host or the dev server, and jekyll just refuses to stay on 127.0.0.1 and writes localhost, which breaks the CPS, and I'm not willing to put localhost logic in there if it's a dev signal. The entire thing can be avoided by just using the root relative path, and that is not a compromised implementation.

This was triply hard because the manifest was being signed based on a blanked path (to avoid the manual derivation of exact strings that were multiple times worse byte-wise than hardcoding magic values). The local site would be happy with this signature, but the deployed site would be checking the signature for a different payload than it's rendering in standard jekyll land.

tl;dr: Don't let jekyll render something different than what we sign