Clean up
Replacing references to the control domain with {{ site.remote }} ensures that deployments are just self-referencing with an empty string, unless they are explicitly remote-controlled. Much simpler, less leaking in the CSP.

Storing the QR in the favicon is pretty funny tbh. That is a hard cache field to purge sometimes.