Fix content-type for specific loads
This is kind of sweeping, but it allows us to map our types to content-type when it matters to the browser for how it loads or triggers internal resources. The most notable example is text/html loaded as an application/manifest+html for the widget system. A strange vendor content-type makes accessing the url a content-disposition:attachment, so iframing it is not possible.